Add more regression tests for btl_enmy_prm
