+++ .gitignore
@@ -1,5 +1,6 @@
 log.txt
 /regression_tests/work_dirs
+/coverage
 
 *.bak
 *.ods
